Prepare for dynamic timesteps
This patch removes the Dynamic.set_timestep method, which is not a good fit in
a situation where time steps may vary. Instead the add_and_plot method is
rewritten to use dimensional times rather than a fixed-length deque, and the
newmark_defo and newmark_velo methods receive a timestep argument to replace
the attribute. The latter will be further modified in a subsequent commit to
allow for truly dynamic time steps.